Use std::unique_ptr. (#853)
* Use std::unique_ptr.

This PR takes advantage of std::unique_ptr added in C++11.  Given that
dynet uses C++11, we can take advantage of the feature for RAII if
applicable.

For the code of dynet library, it would be helpful to clarify the
ownership of objects. In particular, it is useful for debugging memory
related bugs such as memory leaks, use-after-free.

For the code in examples, we can remove the code to manually delete
objects allocated by new, using std::unique_ptr.

* Fix typo


Former-commit-id: 28e3bac77a354eaf837cf9a66b360c3ed9d1cd1e