Eliminate dependency on libdynet from _dynet.so
Currently, the compiled Cython file, _dynet.so depends on libdynet.
This has a couple of disadvantages such as:

- Installation can be clumsy (e.g., setting LD_LIBRARY_PATH in Linux,
  DYLD_LIBRARY_PATH in macOS to load libdynet)
- Not easy to deploy to servers.

This change eliminates the dependency by creating a static library of
dynet, making the installation and deployment easier.  The idea is to
link the static library rather than the shared/dynamic library when
generating _dynet.so.

The static and shared/dynamic libraries are generated from an object
library (it's just a collection of object files) [1].  By creating an
object library, we can avoid compiling object files for both libraries.

[1] https://cmake.org/cmake/help/latest/command/add_library.html#object-libraries