examples: attempt to fix training with standard softmax
BuildLMGraph assumes "cfsm" points to an object
of class factored softmax, but it can point to
an object of StandardSoftmaxBuilder if users run
the command as documented in README.md.

I'm not sure this is a correct fix because the comments
in BuildLMGraph use a bit different term, "regular softmax".
If it refers to the standard softmax, removing allocating
StandardSoftmaxBuilder object seems correct.