Device support in Python.
Currently there is no dedicated Device class, and we refer to devices by
their string names. This seem less cumbersome and more 'Pythonic'.
However, if we forsee a case where we would really want Device to be an
object, we may do this now for backward compatibility.

The API is:
- dy.available_devices() returns a list of device names.
- dy.to_device(expression, dev_name) copies an expression across
  devices.
- param_colection.add_parameteres(..., device=name)
- param_colection.add_lookup_parameteres(..., device=name)
- param_colection.parameteres_from_numpy(..., device=name)
- param_colection.lookup_parameteres_from_numpy(..., device=name)

The devices mechanism documentation is in doc-string of "available_devices()"