TreeLSTMs: add nodes in arbitrary order; swig bindings
Main feature:
TreeLSTMs had to be constructed with the nodes in the tree added by
their index.  This is not practical if the nodes are already
sequentially ordered with the child nodes not being first, e.g. when
working with dependency trees.  If set_num_elements is called, a fixed
set of node space is reserved and nodes can be added in any order as
long as children are added first.

The old behavior still works as usual.

Also:
 - Add documentation to treelstm C++ code
 - add swig bindings
 - add scala wrapper for uni- and bidirectional tree lstm