TreeLSTM additions (#1133)
* make {Nary, Unidirectional}TreeLSTMBuilder non-abstract again

Commit bf29c18fcf1bcc356570f6c4ffd64fec84fd470d added set_h_impl to
RNNBuilder as an abstract method and the TreeLStmBuilders became
abstract.  This was fixed for the BidirectionalTreeLSTMBuilder but not
the other two.  This commit moves the the stub throwing a runtime error
from BidirectionalTreeLSTMBuilder to the TreeLSTMBuilder super class.

* TreeLSTMs: add nodes in arbitrary order; swig bindings

Main feature:
TreeLSTMs had to be constructed with the nodes in the tree added by
their index.  This is not practical if the nodes are already
sequentially ordered with the child nodes not being first, e.g. when
working with dependency trees.  If set_num_elements is called, a fixed
set of node space is reserved and nodes can be added in any order as
long as children are added first.

The old behavior still works as usual.

Also:
 - Add documentation to treelstm C++ code
 - add swig bindings
 - add scala wrapper for uni- and bidirectional tree lstm

* scala uni- and bidirectional tree lstm wrapper